% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{cal_elapsed_days}
\alias{cal_elapsed_days}
\title{Obtain the days from earlier to latter}
\usage{
cal_elapsed_days(earlier, latter)
}
\arguments{
\item{earlier}{the early date}

\item{latter}{the late date}
}
\value{
days in integer
}
\description{
The full name is : calculate elapsed day from `earlier` to `latter`.
}
\examples{
elapsed_days('2021-10-01', '2022-01-01')
}
