% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckData.R
\name{check_anyInf}
\alias{check_anyInf}
\title{Check any Inf values in the data.frame}
\usage{
check_anyInf(df)
}
\arguments{
\item{df}{}
}
\value{
T for has
}
\description{
Check any Inf values in the data.frame
}
\examples{
check_anyInf(df)
}
