% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckData.R
\name{check_anyNA}
\alias{check_anyNA}
\title{Check any NA values in the data.frame}
\usage{
check_anyNA(df)
}
\arguments{
\item{df}{support input automic vector, matrix and data.frame}
}
\value{
T for has
}
\description{
该函数执行的功能为：如果输入的df参数是一个一维向量，那么返回是否有任何一个元素为NA；
如果输入的df参数为一个matrix也判定是否有任一元素为NA；
如果输入的df参数为一个 data.frame 也判定是否有任一行中存在NA
}
\examples{
check_anyNA(df)
}
