% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckData.R
\name{check_anyNaN}
\alias{check_anyNaN}
\title{Check any NaN values in the data.frame}
\usage{
check_anyNaN(df)
}
\arguments{
\item{df}{}
}
\value{
T for has
}
\description{
Check any NaN values in the data.frame
}
\examples{
check_anyNaN(df)
}
