% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckData.R
\name{check_anyNull}
\alias{check_anyNull}
\title{Check any Null values in the data.frame}
\usage{
check_anyNull(df)
}
\arguments{
\item{df}{}
}
\value{
T for has
}
\description{
Check any Null values in the data.frame
}
\examples{
check_anyNull(df)
}
