% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckData.R
\name{check_anyZeroStr}
\alias{check_anyZeroStr}
\title{Check any zero string ("") in the data.frame}
\usage{
check_anyZeroStr(df)
}
\arguments{
\item{df}{}
}
\value{
T for has
}
\description{
Currently only implements the characters vector.
}
\examples{
check_anyZeroStr(df)
}
