% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funCleanDataSet-Josh-st.R
\docType{data}
\name{chr.ref.NC}
\alias{chr.ref.NC}
\title{Table that binds weird Identification to pronounceable name}
\format{An object of class \code{data.frame} with 17 rows and 6 columns.}
\usage{
chr.ref.NC
}
\description{
then used by a useful? function
}
\details{
note: stringAsFactors returns Error. Hum?
therefore, apply as.character to the right columns
}
\keyword{datasets}
