% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{config_ensureMemory}
\alias{config_ensureMemory}
\title{Set the max_size of the memory}
\usage{
config_ensureMemory()
}
\description{
The full name of this function is: configurate the memory for R that ensuring computation.
}
\examples{
config_ensureMemory()
}
