% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioinformaticsKnowledge.R
\name{const_blast_fmt6_defaultVarNames}
\alias{const_blast_fmt6_defaultVarNames}
\title{Constant values of the blast software output format 6 default column names}
\usage{
const_blast_fmt6_defaultVarNames()
}
\value{
a char vector
}
\description{
Constant values of the blast software output format 6 default column names
}
\examples{
colnames(df) <- const_blast_fmt6_defaultVarNames()
}
