% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioinformaticsKnowledge.R
\name{const_hgnc_complete_geneTable}
\alias{const_hgnc_complete_geneTable}
\title{Get human all gene names table.}
\usage{
const_hgnc_complete_geneTable()
}
\value{
a tibble object
}
\description{
Get human all gene names table.
}
\examples{
const_hgnc_complete_geneTable()
}
