% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioinformaticsKnowledge.R
\name{const_hmmer_tblout_full_VarNames}
\alias{const_hmmer_tblout_full_VarNames}
\title{Constant values of the hmmer software tblout.txt format file first 7 column names.}
\usage{
const_hmmer_tblout_full_VarNames()
}
\value{
a char vector
}
\description{
Constant values of the hmmer software tblout.txt format file first 7 column names.
}
\examples{
colnames(df) <- const_hmmer_tblout_full_VarNames()
}
