% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{des_object_str}
\alias{des_object_str}
\title{In R the data structure is too Weird and strange, so we need to take a comprehensive of a variable.}
\usage{
des_object_str(x, complexStr = F)
}
\arguments{
\item{x}{}
}
\value{
No return, just the prints to watch.
}
\description{
In R the data structure is too Weird and strange, so we need to take a comprehensive of a variable.
}
\examples{
des_object_str(x = 1:10)
}
