% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{des_two_charSets}
\alias{des_two_charSets}
\title{Input two char vectors or vector-like variable, describe the content of the two vector.}
\usage{
des_two_charSets(x, y, onlyNum = T)
}
\arguments{
\item{x}{any elements will be convert to the char vector}

\item{y}{any elements will be convert to the char vector}

\item{onlyNum}{show the numbers only}
}
\value{
a list of three elements
}
\description{
The full name of this function is describe two char sets.
}
\examples{
des_two_charSets( 1:10, 2:12)
}
