% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisulizeFunctions.R
\name{draw_ecdfPlot}
\alias{draw_ecdfPlot}
\title{Quick plot the ecdf.}
\usage{
draw_ecdfPlot(data, ...)
}
\arguments{
\item{data}{df with wider format}

\item{...}{parameters to the stat_ecdf}
}
\value{
the ggplot object
}
\description{
Quick plot the ecdf.
}
\examples{
data <- as.data.frame(matrix(rnorm(100 * 5), ncol = 5))
draw_ecdfPlot(data)
}
