% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisulizeFunctions.R
\name{draw_pcpPlot}
\alias{draw_pcpPlot}
\title{Quickly draw the parallel coordinates plots.}
\usage{
draw_pcpPlot(df, ...)
}
\arguments{
\item{df}{the data frame like the demo data mpg in `ggplot2` package.}

\item{...}{the parameters import to the `geom_line()`}
}
\value{
the ggplot2 object
}
\description{
Referenced from the https://ggplot2-book.org/programming#sec-functions.
}
\examples{
draw_pcpPlot(mpg) + theme_bw()
draw_pcpPlot(mpg, aes(colour = drv)) + theme_bw()
}
