% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{export_strList2tmpFile_asTable}
\alias{export_strList2tmpFile_asTable}
\title{Conveniently export the list to temporary file.}
\usage{
export_strList2tmpFile_asTable(l, out_path = NULL)
}
\arguments{
\item{l}{: a list with char as the elements.}
}
\value{
nothing
}
\description{
Conveniently export the list to temporary file.
}
\examples{
export_strList2tmpFile_asTable(list(a = letters, b = 1:10))
}
