% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{force_create_file_evenWithout_dirs}
\alias{force_create_file_evenWithout_dirs}
\title{Force create file even without dirs}
\usage{
force_create_file_evenWithout_dirs(file_path)
}
\arguments{
\item{file_path}{the file path like a/b/c}
}
\value{
nothing
}
\description{
Force create file even without dirs
}
\examples{
force_create_file_evenWithout_dirs(a/b/v)
}
