% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Class_ColorMapper.R
\name{generate_colors}
\alias{generate_colors}
\title{Generate colors for the input x}
\usage{
generate_colors(x, values2colorMapper)
}
\arguments{
\item{x}{matrix or numeric vector}

\item{values2colorMapper}{the instance: discrete or continues}
}
\value{
colors
}
\description{
In fact, I need to admit that: the discrete value to colors mapper, is not practical.
you can directly use the named vector to get the colors corresponding to the value.

targetColors <- RColorBrewer::brewer.pal(9, name = 'Set1')
colorMap <- set_names(targetColors, nm = letters[seq2_along(1,targetColors)])
colorMap[data] |> as.character()
}
\examples{
## This is a low-level situation
a <- discreteMapper(colors = c('red', 'blue'), break_points = c(1,2,3))
generate_colors(1:10,a)

## A real application situation
targetColors <- RColorBrewer::brewer.pal(9, name = 'Set1')
range <- c(1,10);
break_points <- seq.int(from = range[1], to = tail(range,1), length.out = 50)
a <- discreteMapper(colors = colorRampPalette(targetColors)(49) , break_points = break_points)

generate_colors(1:10,a)

}
