% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get_strVector_fromClipboard}
\alias{get_strVector_fromClipboard}
\title{Get the string vector form the clipboard.}
\usage{
get_strVector_fromClipboard()
}
\value{
atomic string vector
}
\description{
Get the string vector form the clipboard.
}
\examples{
get_strVector_fromClipboard()
}
