% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get_vec_element_counts}
\alias{get_vec_element_counts}
\title{Get the element counts of a vector, it use the base::table() function.
Please refer the table function for the suitable situations.}
\usage{
get_vec_element_counts(vec)
}
\arguments{
\item{vec}{a vector, e.g.: atomic vector}
}
\value{
a tibble with two column
}
\description{
Get the element counts of a vector, it use the base::table() function.
Please refer the table function for the suitable situations.
}
\examples{
get_vec_element_counts(letters)
}
