% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiologyUtils.R
\name{getGeneInfoFromPartialGeneNameDesc_mouse}
\alias{getGeneInfoFromPartialGeneNameDesc_mouse}
\title{Get the gene information from partial gene name description from organism mouse.}
\usage{
getGeneInfoFromPartialGeneNameDesc_mouse(str)
}
\arguments{
\item{str}{partial description of the gene. E.g. transcript factor}
}
\value{
the df of the result genes
}
\description{
Get the gene information from partial gene name description from organism mouse.
}
\examples{
getGeneInfoFromPartialGeneNameDesc_mouse('Albumin')
}
