% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getGlobalVars}
\alias{getGlobalVars}
\title{Get the global values.
You may wonder Can we set the global values through the returned env. object?
No. (1) will not persist store in your desk.
(2) Temporary usage is enough setGlobalVars(NULL) to save.}
\usage{
getGlobalVars()
}
\value{
the env to store the values.
}
\description{
Get the global values.
You may wonder Can we set the global values through the returned env. object?
No. (1) will not persist store in your desk.
(2) Temporary usage is enough setGlobalVars(NULL) to save.
}
\examples{
getGlobalVars()
}
