% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funCleanDataSet-Josh-st.R
\name{GetLenghtChr}
\alias{GetLenghtChr}
\title{Return the Length of the chromosome number}
\usage{
GetLenghtChr(chr.num)
}
\arguments{
\item{chr.num}{the integer corresponding to the Chromosome}
}
\description{
Return the Length of the chromosome number
}
