% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funCleanDataSet-Josh-st.R
\name{GetRand}
\alias{GetRand}
\title{Return the Data.Frame with Random Chromosome and Position
(all the weights are set to 1.)}
\usage{
GetRand(nsample)
}
\arguments{
\item{nsample}{the number of samples}
}
\value{
the data frame
}
\description{
Return the Data.Frame with Random Chromosome and Position
(all the weights are set to 1.)
}
\examples{

GetRand(10)

}
