% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funCleanDataSet-Josh-st.R
\name{GetRefChr}
\alias{GetRefChr}
\title{Return the ref|NC_XXXXXX| of any chromosome number}
\usage{
GetRefChr(chr.num)
}
\arguments{
\item{chr.num}{the integer corresponding to the Chromosome}
}
\value{
string of type ref|NC_XXXXXX|
}
\description{
Return the ref|NC_XXXXXX| of any chromosome number
}
