% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funCleanDataSet-Josh-st.R
\name{GetSubtelo}
\alias{GetSubtelo}
\title{Return the Data.Frame updated if the position belong to the Subtelo region}
\usage{
GetSubtelo(df)
}
\arguments{
\item{df}{the Data.Frame returned by GetWeight}
}
\value{
the data frame df updated
}
\description{
Return the Data.Frame updated if the position belong to the Subtelo region
}
