% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funCleanDataSet-Josh-st.R
\name{GetTinySample}
\alias{GetTinySample}
\title{Randomly select n rows from data.frame}
\usage{
GetTinySample(df, n = 20)
}
\arguments{
\item{df}{the data.frame}

\item{n}{the number of randomly selected rows (default: 20)}
}
\value{
a subset and tiny data.frame
}
\description{
Randomly select n rows from data.frame
}
