% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funCleanDataSet-Josh-st.R
\name{Is.InSubtelo}
\alias{Is.InSubtelo}
\title{Return the boolean of Subtelo region}
\usage{
Is.InSubtelo(pos, chrX, length.subtelo = 30000)
}
\arguments{
\item{pos}{the position (integer)}

\item{chrX}{the string of the chromosome}

\item{length.subtelo}{by default 30000}
}
\description{
Return the boolean of Subtelo region
}
