% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statstics.R
\name{jaccard_coefficient}
\alias{jaccard_coefficient}
\title{Function to calculate the Jaccard coefficient
Input: Two character vectors
Output: Jaccard coefficient}
\usage{
jaccard_coefficient(set1, set2)
}
\arguments{
\item{vec1}{char vector1}

\item{vec2}{char vector2}
}
\value{
the jcc
}
\description{
Function to calculate the Jaccard coefficient
Input: Two character vectors
Output: Jaccard coefficient
}
\examples{
jaccard_coefficient(letters[1:10], letters[5:12])
vec1 <- c("apple", "banana", "cherry")
vec2 <- c("banana", "cherry", "date", "apple")
jaccard_index <- jaccard_coefficient(vec1, vec2)
print(jaccard_index)
}
