% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funCleanDataSet-Josh-st.R
\name{MergeOnly}
\alias{MergeOnly}
\title{Merge only one chromosome from two data.frames
      (?not useful?)}
\usage{
MergeOnly(df1, df2, kro = "chrXX")
}
\arguments{
\item{df1}{first data.frame}

\item{df2}{second data.frame}

\item{kro}{name of one chromosome}
}
\value{
the merged data.frame
}
\description{
The data.frames come from the function GetWeight
}
\examples{

df1 comes from AB15
df2 comes from AB16

kro is "chrI" or "ref|NC_001133|"

}
