% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{orgnize_strList2table}
\alias{orgnize_strList2table}
\title{Orgnize the string list to the table(tibble)}
\usage{
orgnize_strList2table(l)
}
\arguments{
\item{l}{most be a 'list' and must has name attributes}
}
\value{
a table with row number is max element length, padding with ""
}
\description{
Orgnize the string list to the table(tibble)
}
\examples{
orgnize_strList2table(list(a = letters[1:5], b = LETTERS[1:3], c = c('gg')))
}
