% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{pas_pathPrep_MSWindows}
\alias{pas_pathPrep_MSWindows}
\title{Prepare the path for windows. If no argument input, it will read the content from the clipboard.
It not only return the value, but change the clip board.}
\usage{
pas_pathPrep_MSWindows(path = "clipboard")
}
\arguments{
\item{path:}{if no input, it will read content from the clipboard.}
}
\value{
: path string represents the path in ms windows.
}
\description{
This is a paste function, very convenient for Windows user.
The full name is: paste file path which is prepared for MS Windows.
}
\examples{
pathPrep_MSWindows()
}
