% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{pas_vectorElements2str}
\alias{pas_vectorElements2str}
\title{Put string vector elements to string.}
\usage{
pas_vectorElements2str(v)
}
\arguments{
\item{v}{a vecter, better to be char vector}
}
\value{
the string of assign a vector
}
\description{
The full name of this function is: paste the vector element to string that for assignment.
}
\examples{
# 创建一个字符向量 v
v <- c("a", "b", "c")
使用 paste 函数将字符向量打印成字符串格式
output <- paste("c(", paste0('"', v, '"', collapse = ","), ")")
# 输出结果
cat(output)
另一个有趣的例子：
tryR::pas_vectorElements2str(1:10) |> cat()
}
