% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisulizeFunctions.R
\name{piechart}
\alias{piechart}
\title{Quickly plot a piechart of a plot.}
\usage{
piechart(data, mapping)
}
\arguments{
\item{data}{the long format data frame.}

\item{mapping}{the mapping object of ggplot2}
}
\value{
the ggplot object
}
\description{
https://ggplot2-book.org/programming#sec-functions
}
\examples{
piechart(mpg, aes(factor(1), fill = drv)) + theme_bw()
}
