% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisulizeFunctions.R
\name{plot_points_sequential}
\alias{plot_points_sequential}
\title{Plot points sequential.}
\usage{
plot_points_sequential(x, y, default.units = "in")
}
\arguments{
\item{x}{x coordinates}

\item{y}{y coordinates}

\item{default.units}{default is inch}
}
\description{
Plot points sequential.
}
\examples{

plot_points_sequential(x = 1:3, y = 1:3)
}
