% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funCleanDataSet-Josh-st.R
\name{rBindOnly.Signed}
\alias{rBindOnly.Signed}
\title{Bind by row only one chromosome from two data.frames}
\usage{
rBindOnly.Signed(df1, df2, kro = "chrXX", name1 = "One", name2 = "Two")
}
\arguments{
\item{df1}{first data.frame}

\item{df2}{second data.frame}

\item{kro}{name of one chromosome}

\item{name1}{associates a name to df1}

\item{name2}{associates a name to df2}
}
\value{
the binded data.frame by row
}
\description{
The data.frames come from the function GetWeight
}
\examples{

df1 comes from AB15
df2 comes from AB16

kro is "chrI" or "ref|NC_001133|"

}
