% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funCleanDataSet-Josh-st.R
\name{Ref2Str}
\alias{Ref2Str}
\title{Return the chrX of any chromosome reference id (ref|NC_XXXXXX|)}
\usage{
Ref2Str(ref)
}
\arguments{
\item{ref}{the string corresponding to Refernce id: ref|NC_XXXXXX|}
}
\value{
string of type chrX
}
\description{
Return the chrX of any chromosome reference id (ref|NC_XXXXXX|)
}
