% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{reorder_rows}
\alias{reorder_rows}
\title{定义一个自定义函数来进行排序}
\usage{
reorder_rows(data, target_names, column_name)
}
\arguments{
\item{data}{a tibble}

\item{target_names}{the vector for filter rows}

\item{column_name}{the column name to filter}
}
\value{
a filtered tibble
}
\description{
定义一个自定义函数来进行排序
}
\examples{
reorder_rows(tibble, c('a','b','c'), col1)
}
