% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{rm_temp_vars}
\alias{rm_temp_vars}
\title{Clear Temporary variables}
\usage{
rm_temp_vars(remove_pattern = "^tmp_")
}
\arguments{
\item{remove_pattern}{The default pattern remove the header string tmp}
}
\description{
Clear Temporary variables
}
\examples{
rm_temp_vars()
}
