% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{setGlobalVars}
\alias{setGlobalVars}
\title{set The global Vars.
Current support is:}
\usage{
setGlobalVars(varList)
}
\arguments{
\item{varList}{a named vector with string key-values.}
}
\value{
no return
}
\description{
hgnc_complete_genes_path
tmpFile
}
\examples{
setGlobalVars(NULL); # Just for storage
setGlobalVars( list( tmpFile= 'path', hgnc_complete_genes_path = '/your/path/file.txt') )
}
