% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisulizeFunctions.R
\name{transNumericDat2longFormat}
\alias{transNumericDat2longFormat}
\title{Transform the data.frame to be the scaled values. Only apply to the numeric values.}
\usage{
transNumericDat2longFormat(df)
}
\arguments{
\item{df}{data.frame class instance}
}
\value{
scaled data.frame with long format. The numeric columns are substitute to variable and value columns.
}
\description{
Transform the data.frame to be the scaled values. Only apply to the numeric values.
}
\examples{
pce_data(mpg)
}
