% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statstics.R
\name{two_groups_correlationTest}
\alias{two_groups_correlationTest}
\title{Do the hypothesis testing for two correlation coefficient groups.}
\usage{
two_groups_correlationTest(x = c(0.1, 0.2, 0.3), y = c(0.4, 0.5, 0.6))
}
\arguments{
\item{x}{the pcc of group1}

\item{y}{the pcc of group2}
}
\value{
the result of the t test
}
\description{
Reference: https://blog.csdn.net/CoolCoolCarrot/article/details/106001211
}
\examples{
two_groups_correlationTest()
}
