% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ty-package.R
\docType{package}
\name{ty-package}
\alias{ty}
\alias{ty-package}
\title{ty: The convenient Utility functions designed for Dalang Yu’s personal use.}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
