% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisulizeFunctions.R
\name{venn_plot}
\alias{venn_plot}
\title{Draw a venn graph of three sets with good outlook, the dark theme.}
\usage{
venn_plot(x, ...)
}
\arguments{
\item{x}{a list contains tree sets}
}
\value{
the intersections and diff results.
}
\description{
Draw a venn graph of three sets with good outlook, the dark theme.
}
\examples{
plotThreeSetsVennGraph_deep()
}
