% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LightWeightPlot.R
\name{yyBoxplot}
\alias{yyBoxplot}
\title{Quick plot the box plot}
\usage{
yyBoxplot(x, colo = NULL)
}
\arguments{
\item{x}{: A list with names and values}

\item{colo}{: colors}
}
\value{
no values
}
\description{
Quick plot the box plot
}
\examples{
x <- list(A = 1:10, B = 2:11, C = 0:9)
colors <- RColorBrewer::brewer.pal(n = 3, "Paired")
yyBoxplot(x,colors)

}
