% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funCleanDataSet-Josh-st.R
\name{GetWeight}
\alias{GetWeight}
\title{Read one file and return the data.frame}
\usage{
GetWeight(mypath)
}
\arguments{
\item{mypath}{the path of the file}
}
\value{
the data.frame with the columns: chr pos orient weight

Need to be a bit more cleanned !! :-)
(some stuff are not useful anymore)
}
\description{
Read one file and return the data.frame
}
\examples{

AB15 <- system.file("extdata", "AB15.light.txt", package = "Ty")
AB16 <- system.file("extdata", "AB16.light.txt", package = "Ty")

r15 <- GetWeight(AB15)
r16 <- GetWeight(AB16)


}
